% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixer-package.R
\docType{package}
\name{fixer-package}
\alias{fixer}
\alias{fixer-package}
\title{fixer: Parse Fixed-Width Files According to their Dictionary}
\description{
Some fixed-width files (fwf) are created with a complementary data dictionary. Such dictionaries contain metadata such as column widths, variable types, and variable labels. The fixer package reads fixed-width files according to their data dictionary.
}
\author{
\strong{Maintainer}: Travis Gerke \email{tgerke@mail.harvard.edu} (\href{https://orcid.org/0000-0002-9500-8907}{ORCID})

Authors:
\itemize{
  \item Konrad Stopsack \email{stopsack@post.harvard.edu} (\href{https://orcid.org/0000-0002-0722-1311}{ORCID})
}

}
\keyword{internal}
