% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dd.R
\name{read_dd}
\alias{read_dd}
\title{Read a fwf data dictionary file}
\usage{
read_dd(path, skip = 0)
}
\arguments{
\item{path}{Path to the data dictionary file}

\item{skip}{Number of lines to skip before reading data}
}
\value{
Arguments to readr::read_fwf and labelled::set_variable_labels
}
\description{
Makes some assumptions about the structure of the data dictionary file. (I
think, but TODO let's describe these if that ends up being true).
}
\examples{
fixer:::read_fwf_example("example-fwf-dictionary.dd") |>
  read_dd(skip = 2)
}
