% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fwf_dd.R
\name{read_fwf_dd}
\alias{read_fwf_dd}
\title{Read a fixed-width file when given a dictionary}
\usage{
read_fwf_dd(path_fwf, dd, unlabel = TRUE)
}
\arguments{
\item{path_fwf}{Path to the fixed-width file}

\item{dd}{A data dictionary object, typically defined by read_dd()}

\item{unlabel}{Logical, when TRUE factor codes are replaced with the labels}
}
\value{
A tibble
}
\description{
Read a fixed-width file when given a dictionary
}
\examples{
dd <- fixer:::read_fwf_example("example-fwf-dictionary.dd") |>
  read_dd(skip = 2)
fixer:::read_fwf_example("example-fwf-data.dat") |>
  read_fwf_dd(dd)
}
